Add landing page.

+++ database/migrations/2020_06_14_083813_create_invoice_items_table.php
@@ -18,7 +18,7 @@ class CreateInvoiceItemsTable extends Migration
             $table->bigInteger('invoice_id');
             $table->bigInteger('item_quantity');
             $table->string('item_description');
-            $table->float('item_cost');
+            $table->double('item_cost', 15, 8);
             $table->timestamps();
         });
     }
